Minor code improvements

+++ 3rdparty/indi-asi/asi_ccd.h
@@ -145,8 +145,8 @@ class ASICCD : public INDI::CCD
     ITextVectorProperty SDKVersionSP;
 
     struct timeval ExpStart;
-    float ExposureRequest;
-    float TemperatureRequest;
+    double ExposureRequest;
+    double TemperatureRequest;
 
     ASI_CAMERA_INFO *m_camInfo;
     ASI_CONTROL_CAPS *pControlCaps;
